% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_analysis.R
\name{apply_conditional_formatting}
\alias{apply_conditional_formatting}
\title{Apply Conditional Formatting}
\usage{
apply_conditional_formatting(filename, p_value_column)
}
\arguments{
\item{filename}{Path to the Excel file.}

\item{p_value_column}{Column letter for p-values to format.}
}
\description{
Applies conditional formatting to an Excel file based on p-values.
}
